adds adding/removing selections partially by holding shift/ctrl
* renames KeyManager and rewrites it
* some refactoring in selection control

+++ src/glwidget.h
@@ -56,8 +56,8 @@ private:
     QVector3D mCameraDir;
     QPointF mMousePos;
     QPointF mAngle; // (x, y) camera angle in degree
-    qreal mMouseSensitivy;
-    qreal mMovementSpeed;
+    float mMouseSensitivy;
+    float mMovementSpeed;
 
     Renderer* mRenderer;
     UIControl* mUiControl;
